FP Kriptografi/project-folder/requirements.txt
Flask
pycryptodome
numpy
